+++ packages/keyv/test/keyv-test.test.ts
@@ -3,7 +3,6 @@ import {
 } from 'vitest';
 import {faker} from '@faker-js/faker';
 import {Keyv} from '../src/index.js';
-import {createKeyv} from '../src/generic-store.js';
 
 describe('Keyv', async () => {
 	type TestData = {
@@ -64,105 +63,23 @@ describe('Keyv', async () => {
 			const resultValue = await keyv.get(testData[0].key);
 			expect(resultValue).toEqual(testData[0].value);
 		});
-
-		test('should use the store to set multiple keys', async () => {
-			const map = new Map();
-			const keyv = createKeyv(map);
-			const result = await keyv.setMany(testData);
-			expect(result).toEqual([true, true, true, true, true]);
-			const resultValue = await keyv.get(testData[0].key);
-			expect(resultValue.value).toEqual(testData[0].value);
-		});
-
-		test('should emit and return false on error', async () => {
-			const map = new Map();
-			map.set = () => {
-				throw new Error('Test Error');
-			};
-
-			const keyv = createKeyv(map);
-			let errorEmitted = false;
-			keyv.on('error', () => {
-				errorEmitted = true;
-			});
-
-			const result = await keyv.setMany(testData);
-			expect(result).toEqual([false, false, false, false, false]);
-			expect(errorEmitted).toBe(true);
-		});
 	});
 
-	describe('deleteMany', async () => {
-		test('should emit and return false on error', async () => {
-			const map = new Map();
-			const keyv = createKeyv(map);
-			keyv.store.deleteMany = () => {
-				throw new Error('Test Error');
-			};
-
-			let errorEmitted = false;
-			keyv.on('error', () => {
-				errorEmitted = true;
-			});
-
-			const result = await keyv.deleteMany(testKeys);
-			expect(result).toEqual(false);
-			expect(errorEmitted).toBe(true);
-		});
-	});
-
-	describe('getMany', async () => {
-		test('should set many items and then get them', async () => {
-			const keyv = createKeyv(new Map());
-			await keyv.setMany(testData);
-			const result = await keyv.getMany(testKeys);
-			expect(result.length).toBe(5);
+	describe('throwErrors', async () => {
+		test('should get the current throwErrors value', async () => {
+			const keyv = new Keyv(new Map());
+			expect(keyv.throwErrors).toBe(false);
 		});
 
-		test('should set many items and then get them with get', async () => {
-			const keyv = createKeyv(new Map());
-			await keyv.setMany(testData);
-			const result = await keyv.get(testKeys);
-			expect(result.length).toBe(5);
+		test('should set the throwErrors value', async () => {
+			const keyv = new Keyv(new Map());
+			keyv.throwErrors = true;
+			expect(keyv.throwErrors).toBe(true);
 		});
 
-		test('should set many items and then get them raw', async () => {
-			const keyv = createKeyv(new Map());
-			await keyv.setMany(testData);
-			const result = await keyv.getMany(testKeys, {raw: true});
-			expect(result.length).toBe(5);
-			expect(result[0]?.value.value).toBe(testData[0].value);
-		});
-	});
-
-	describe('hasMany', async () => {
-		test('should set many items and then check if they exist', async () => {
-			const keyv = createKeyv(new Map());
-			await keyv.setMany(testData);
-			const result = await keyv.hasMany(testKeys);
-			expect(result.length).toBe(5);
-		});
-
-		test('should use the store hasMany function', async () => {
-			const map = new Map();
-			const keyv = createKeyv(map);
-			keyv.store.hasMany = async () => [true, true, true, true, true];
-
-			await keyv.setMany(testData);
-			const result = await keyv.has(testKeys);
-			expect(result.length).toBe(5);
-		});
-
-		test('should be able to get less on hasMany', async () => {
-			const keyv = createKeyv(new Map());
-			await keyv.setMany(testData);
-			const resultList = await keyv.hasMany(testKeys);
-			expect(resultList.length).toBe(5);
-			const deleteResult = await keyv.delete(testData[0].key);
-			expect(deleteResult).toBe(true);
-			const result = await keyv.hasMany(testKeys);
-			expect(result.length).toBe(5);
-			expect(result[0]).toBe(false);
+		test('should pass in the throwErrors option', async () => {
+			const keyv = new Keyv(new Map(), {throwErrors: true});
+			expect(keyv.throwErrors).toBe(true);
 		});
 	});
 });
